+++ src/main/java/com/williamhsieh/fruitandessence/model/Product.java
@@ -12,7 +12,7 @@ public class Product {
     private ProductCategory category;
     private String imageUrl;
 
-    private Integer stock; // 庫存量
+    private Double stock; // 庫存量
     private Integer pricePerUnit; // 每單位價格
     private String unit; // 單位
     private ProductUnitType unitType; // WEIGHT 或 COUNT
@@ -55,11 +55,11 @@ public class Product {
         this.imageUrl = imageUrl;
     }
 
-    public Integer getStock() {
+    public Double getStock() {
         return stock;
     }
 
-    public void setStock(Integer stock) {
+    public void setStock(Double stock) {
         this.stock = stock;
     }
